Got a basic Diesel query for Part D working.
It's not _really_ used anywhere, and I'm not even _really_ verifying
that it works... But. All the machinery is there, and I _think_ it's
working as expected.

This is really just a good point to pause and save a snapshot, before I
move on to actually trying to really use this stuff.